1. Code linting 2. Rename pumpkinreader.rest to pumpkinreader.service
